finally all aspects ok for triple key repo

+++ SharpRepository.Repository/CompoundKeyRepositoryBase.cs
@@ -903,7 +903,12 @@ namespace SharpRepository.Repository
         {
             if (entity == null) throw new ArgumentNullException("entity");
 
+            if (!RunAspect(attribute => attribute.OnUpdateExecuting(entity, _repositoryActionContext)))
+                return;
+
             UpdateItem(entity);
+
+            RunAspect(attribute => attribute.OnUpdateExecuted(entity, _repositoryActionContext));
             if (BatchMode) return;
 
             Save();
@@ -914,9 +919,14 @@ namespace SharpRepository.Repository
 
         private void Save()
         {
+            if (!RunAspect(attribute => attribute.OnSaveExecuting(_repositoryActionContext)))
+                return;
+
             SaveChanges();
 
             _queryManager.OnSaveExecuted();
+
+            RunAspect(attribute => attribute.OnSaveExecuted(_repositoryActionContext));
         }
 
         protected virtual bool GetPrimaryKey(T entity, out TKey key, out TKey2 key2)
@@ -985,6 +995,8 @@ namespace SharpRepository.Repository
         // the query manager uses the caching strategy to determine if it should check the cache or run the query
         private CompoundKeyQueryManager<T, TKey, TKey2, TKey3> _queryManager;
 
+        private readonly RepositoryActionContext<T, TKey,TKey2,TKey3> _repositoryActionContext;
+
         public override bool CacheUsed
         {
             get { return _queryManager.CacheUsed; }
@@ -1010,6 +1022,9 @@ namespace SharpRepository.Repository
             }
 
             CachingStrategy = cachingStrategy ?? new NoCachingStrategy<T, TKey, TKey2, TKey3>();
+
+            _repositoryActionContext = new RepositoryActionContext<T, TKey, TKey2, TKey3>(this);
+            RunAspect(aspect => aspect.OnInitialized(_repositoryActionContext));
         }
 
         public ICompoundKeyCachingStrategy<T, TKey, TKey2, TKey3> CachingStrategy
@@ -1067,47 +1082,116 @@ namespace SharpRepository.Repository
 
         public override IEnumerable<TResult> GetAll<TResult>(Expression<Func<T, TResult>> selector, IQueryOptions<T> queryOptions, IFetchStrategy<T> fetchStrategy)
         {
-            if (selector == null) throw new ArgumentNullException("selector");
+            try
+            {
+                if (selector == null) throw new ArgumentNullException("selector");
 
-            return _queryManager.ExecuteGetAll(
-                () => GetAllQuery(queryOptions, fetchStrategy).Select(selector).ToList(),
-                selector,
-                queryOptions
-                );
+                var context = new CompoundTripleKeyRepositoryQueryMultipleContext<T, TKey, TKey2, TKey3, TResult>(this, null, queryOptions, selector);
+                if (!RunAspect(attribute => attribute.OnGetAllExecuting(context)))
+                    return context.Results;
+
+                // if the aspect altered the specificaiton then we need to run a FindAll with that specification
+                IEnumerable<TResult> results;
+
+                if (context.Specification == null)
+                {
+                    results = _queryManager.ExecuteGetAll(
+                        () => GetAllQuery(context.QueryOptions, fetchStrategy).Select(context.Selector).ToList(),
+                        context.Selector,
+                        context.QueryOptions
+                        );
+                }
+                else
+                {
+                    context.Specification.FetchStrategy = fetchStrategy;
+
+                    results = _queryManager.ExecuteFindAll(
+                        () => FindAllQuery(context.Specification, context.QueryOptions).Select(context.Selector).ToList(),
+                        context.Specification,
+                        context.Selector,
+                        context.QueryOptions
+                        );
+                }
+
+                context.Results = results;
+                RunAspect(attribute => attribute.OnGetAllExecuted(context));
+
+                return context.Results;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         protected abstract T GetQuery(TKey key, TKey2 key2, TKey3 key3);
 
         public T Get(TKey key, TKey2 key2, TKey3 key3)
         {
-            return _queryManager.ExecuteGet(
-                () => GetQuery(key, key2, key3),
-                null,
-                key,
-                key2,
-                key3
-                );
+            try
+            {
+                var context = new CompoundTripleKeyRepositoryGetContext<T, TKey, TKey2, TKey3>(this, key, key2, key3);
+                if (!RunAspect(attribute => attribute.OnGetExecuting(context)))
+                    return context.Result;
+
+                var result = _queryManager.ExecuteGet(
+                        () => GetQuery(context.Id, context.Id2, context.Id3),
+                        context.Selector,
+                        context.Id,
+                        context.Id2,
+                        context.Id3
+                    );
+
+                context.Result = result;
+                RunAspect(attribute => attribute.OnGetExecuted(context));
+
+                return context.Result;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         public TResult Get<TResult>(TKey key, TKey2 key2, TKey3 key3, Expression<Func<T, TResult>> selector)
         {
-            if (selector == null) throw new ArgumentNullException("selector");
+            try
+            {
+                if (selector == null) throw new ArgumentNullException("selector");
 
-            return _queryManager.ExecuteGet(
-                () =>
-                {
-                    var result = GetQuery(key, key2, key3);
-                    if (result == null)
-                        return default(TResult);
+                var context = new CompoundTripleKeyRepositoryGetContext<T, TKey, TKey2, TKey3, TResult>(this, key, key2, key3, selector);
+                if (!RunAspect(attribute => attribute.OnGetExecuting(context)))
+                    return context.Result;
 
-                    var results = new[] { result };
-                    return results.AsQueryable().Select(selector).First();
-                },
-                selector,
-                key,
-                key2,
-                key3
-                );
+                // get the full entity, possibly from cache
+                var item = _queryManager.ExecuteGet(
+                    () =>
+                    {
+                        var result = GetQuery(key, key2, key3);
+                        if (result == null)
+                            return default(TResult);
+
+                        var results = new[] { result };
+                        return results.AsQueryable().Select(selector).First();
+                    },
+                    selector,
+                    context.Id,
+                    context.Id2,
+                    context.Id3
+                    );
+
+                context.Result = item;
+                RunAspect(attribute => attribute.OnGetExecuted(context));
+
+                return context.Result;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         public bool Exists(TKey key, TKey2 key2, TKey3 key3)
@@ -1147,38 +1231,91 @@ namespace SharpRepository.Repository
 
         public override IEnumerable<T> FindAll(ISpecification<T> criteria, IQueryOptions<T> queryOptions = null)
         {
-            if (criteria == null) throw new ArgumentNullException("criteria");
+            if (criteria == null) return GetAll(queryOptions);
 
-            return _queryManager.ExecuteFindAll(
-                () => FindAllQuery(criteria, queryOptions).ToList(),
-                criteria,
-                null,
-                queryOptions
-                );
+            try
+            {
+                var context = new CompoundTripleKeyRepositoryQueryMultipleContext<T, TKey, TKey2, TKey3>(this, criteria, queryOptions);
+                if (!RunAspect(attribute => attribute.OnFindAllExecuting(context)))
+                    return context.Results;
+
+                var results = _queryManager.ExecuteFindAll(
+                    () => FindAllQuery(context.Specification, context.QueryOptions).ToList(),
+                    context.Specification,
+                    null,
+                    context.QueryOptions
+                    );
+
+                context.Results = results;
+                RunAspect(attribute => attribute.OnFindAllExecuted(context));
+
+                return context.Results;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         public override IEnumerable<TResult> FindAll<TResult>(ISpecification<T> criteria, Expression<Func<T, TResult>> selector, IQueryOptions<T> queryOptions = null)
         {
-            if (criteria == null) throw new ArgumentNullException("criteria");
+            if (criteria == null) return GetAll(selector, queryOptions);
 
-            return _queryManager.ExecuteFindAll(
-                () => FindAllQuery(criteria, queryOptions).Select(selector).ToList(),
-                criteria,
-                selector,
-                queryOptions
-                );
+            try
+            {
+                if (selector == null) throw new ArgumentNullException("selector");
+
+                var context = new CompoundTripleKeyRepositoryQueryMultipleContext<T, TKey, TKey2, TKey3, TResult>(this, criteria, queryOptions, selector);
+                if (!RunAspect(attribute => attribute.OnFindAllExecuting(context)))
+                    return context.Results;
+
+                var results = _queryManager.ExecuteFindAll(
+                    () => FindAllQuery(context.Specification, context.QueryOptions).Select(context.Selector).ToList(),
+                    context.Specification,
+                    context.Selector,
+                    context.QueryOptions
+                    );
+
+                context.Results = results;
+                RunAspect(attribute => attribute.OnFindAllExecuted(context));
+
+                return context.Results;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         public override T Find(ISpecification<T> criteria, IQueryOptions<T> queryOptions = null)
         {
-            if (criteria == null) throw new ArgumentNullException("criteria");
+            try
+            {
+                if (criteria == null) throw new ArgumentNullException("criteria");
 
-            return _queryManager.ExecuteFind(
-                () => FindQuery(criteria, queryOptions),
-                criteria,
-                null,
-                null
-                );
+                var context = new CompoundTripleKeyRepositoryQuerySingleContext<T, TKey, TKey2, TKey3>(this, criteria, queryOptions);
+                if (!RunAspect(attribute => attribute.OnFindExecuting(context)))
+                    return context.Result;
+
+                var item = _queryManager.ExecuteFind(
+                    () => FindQuery(context.Specification, context.QueryOptions),
+                    context.Specification,
+                    null,
+                    null
+                    );
+
+                context.Result = item;
+                RunAspect(attribute => attribute.OnFindExecuted(context));
+
+                return context.Result;
+            }
+            catch (Exception ex)
+            {
+                Error(ex);
+                throw;
+            }
         }
 
         public override TResult Find<TResult>(ISpecification<T> criteria, Expression<Func<T, TResult>> selector, IQueryOptions<T> queryOptions = null)
@@ -1224,7 +1361,12 @@ namespace SharpRepository.Repository
         {
             if (entity == null) throw new ArgumentNullException("entity");
 
+            if (!RunAspect(attribute => attribute.OnAddExecuting(entity, _repositoryActionContext)))
+                return;
+
             AddItem(entity);
+
+            RunAspect(attribute => attribute.OnAddExecuted(entity, _repositoryActionContext));
             if (BatchMode) return;
 
             Save();
@@ -1237,7 +1379,12 @@ namespace SharpRepository.Repository
         {
             if (entity == null) throw new ArgumentNullException("entity");
 
+            if (!RunAspect(attribute => attribute.OnDeleteExecuting(entity, _repositoryActionContext)))
+                return;
+
             DeleteItem(entity);
+
+            RunAspect(attribute => attribute.OnDeleteExecuted(entity, _repositoryActionContext));
             if (BatchMode) return;
 
             Save();
@@ -1259,7 +1406,12 @@ namespace SharpRepository.Repository
         {
             if (entity == null) throw new ArgumentNullException("entity");
 
+            if (!RunAspect(attribute => attribute.OnUpdateExecuting(entity, _repositoryActionContext)))
+                return;
+
             UpdateItem(entity);
+
+            RunAspect(attribute => attribute.OnUpdateExecuted(entity, _repositoryActionContext));
             if (BatchMode) return;
 
             Save();
@@ -1270,9 +1422,13 @@ namespace SharpRepository.Repository
 
         private void Save()
         {
+            if (!RunAspect(attribute => attribute.OnSaveExecuting(_repositoryActionContext)))
+                return;
+
             SaveChanges();
 
             _queryManager.OnSaveExecuted();
+            RunAspect(attribute => attribute.OnSaveExecuted(_repositoryActionContext));
         }
 
         protected virtual bool GetPrimaryKey(T entity, out TKey key, out TKey2 key2, out TKey3 key3)
